fix: resolve clippy warnings across V2 pipeline crate
Fix iterator patterns (needless_range_loop, manual_memcpy), dead code
(cache hash field), matches! macro usage, and excessive float precision.

+++ crates/rasmcore-pipeline-v2/src/filters/enhancement.rs
@@ -1667,7 +1667,7 @@ impl GpuFilter for Clahe {
     fn shader_body(&self) -> &str { enh_shaders::CLAHE_APPLY }
     fn workgroup_size(&self) -> [u32; 3] { [256, 1, 1] }
     fn params(&self, w: u32, h: u32) -> Vec<u8> {
-        let grid = self.tile_grid as u32;
+        let grid = self.tile_grid;
         let tile_w = w / grid.max(1);
         let tile_h = h / grid.max(1);
         let mut buf = Vec::with_capacity(32);
